combine team glob validation into mapper
