+++ lib/code_ownership/private/ownership_mappers/team_globs.rb
@@ -8,6 +8,7 @@ module CodeOwnership
       class TeamGlobs
         extend T::Sig
         include Interface
+        include Validations::Interface
 
         @@map_files_to_owners = T.let(@map_files_to_owners, T.nilable(T::Hash[String, T.nilable(::CodeTeams::Team)])) # rubocop:disable Style/ClassVars
         @@map_files_to_owners = {} # rubocop:disable Style/ClassVars
@@ -113,6 +114,23 @@ module CodeOwnership
         sig { override.returns(String) }
         def description
           'Team-specific owned globs'
+        end
+
+        sig { override.params(files: T::Array[String], autocorrect: T::Boolean, stage_changes: T::Boolean).returns(T::Array[String]) }
+        def validation_errors(files:, autocorrect: true, stage_changes: true)
+          overlapping_globs = OwnershipMappers::TeamGlobs.new.find_overlapping_globs
+
+          errors = T.let([], T::Array[String])
+
+          if overlapping_globs.any?
+            errors << <<~MSG
+              `owned_globs` cannot overlap between teams. The following globs overlap:
+
+              #{overlapping_globs.map { |overlap| "- #{overlap.description}"}.join("\n")}
+            MSG
+          end
+
+          errors
         end
       end
     end
